backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15
django-filter>=24.0
django-cors-headers>=4.0
psycopg2-binary>=2.9
google-generativeai>=0.8.0
gunicorn>=22.0
